only one unit test for make_module
