--- test/parameter.cpp
@@ -1,47 +1,0 @@
-#include <disposer/core/make_module.hpp>
-
-#define BOOST_TEST_MODULE dimension solve
-#include <boost/test/included/unit_test.hpp>
-
-
-using namespace disposer;
-
-using namespace hana::literals;
-using namespace disposer::literals;
-using namespace std::literals::string_view_literals;
-
-
-using type_index = boost::typeindex::ctti_type_index;
-
-template < std::size_t D >
-using ic = index_component< D >;
-
-
-BOOST_AUTO_TEST_CASE(test_1){
-	constexpr auto list = dimension_list{dimension_c< double, char, float >};
-
-	struct exec{
-		constexpr void operator()()const{}
-	};
-
-	constexpr module_init_fn state_dummy{};
-	constexpr exec_fn exec_dummy{exec{}};
-
-	module_make_data module_data{};
-	auto const module_base_ptr = make_module_ptr(list,
-		module_configure{
-			set_dimension_fn([](auto const&){
-				return solved_dimensions{index_component< 0 >{2}};
-			}),
-			make("p1"_param, type_ref_c< 0 >, default_value(3.14159f))
-		}, module_data, state_dummy, exec_dummy);
-	auto const module_ptr = dynamic_cast< module<
-			type_list< float >,
-			hana::tuple<>,
-			hana::tuple<>,
-			hana::tuple< parameter< decltype("p1"_param), float > >,
-			void,
-			exec
-		>* >(module_base_ptr.get());
-	BOOST_TEST(module_ptr != nullptr);
-}
